Fixed product inventory to use number of skids instead of pieces

+++ CWBDataApp/models.py
@@ -249,7 +249,7 @@ class Productinventory(models.Model):
     colour = models.CharField(db_column='colour', max_length=15)
     embossed = models.BooleanField(db_column='Embossed')
     doublesided = models.BooleanField(db_column='DoubleSided')
-    numberofskids = models.SmallIntegerField(db_column='NumberOfSkids')
+    numberofskids = models.DecimalField(db_column='NumberOfSkids', max_digits=7, decimal_places=1)
 
     class Meta:
         managed = False
